backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
httpx
python-dotenv
nltk
aiofiles
typing_extensions
python-multipart
google-generativeai